refactor: consolidate modal logic into ModalManager class
- Merged scattered modal-related functions into a new `ModalManager` class for centralization and my own sanity.
- Replaced brittle imports and spaghetti logic from `modal.js` and `modal-events.js` with cleaner methods in `ModalManager`.
- Adjusted event handling to use `ModalManager` for consistency and have fewer sleepless nights.
- Updated `initializeModal`, `openModal`, and related calls in `index.js`.

Tech debt:
- Still married to jQuery, and no, a divorce isn't happening anytime soon.
- Some duplication in dropdown handling that screams for a second refactor pass (aka, future-you's problem).

Gotchas:
- Anyone touching modal-related functions from the deleted files will get a lovely surprise—they're gone.
- Poppers initialization isn't particularly robust. Pray no one breaks it.
