+++ scripts/events/characters-events.js
@@ -10,7 +10,6 @@ import {
     toggleFavoriteStatus,
     update_avatar,
 } from '../components/characters.js';
-import { closeDetails } from '../components/modal.js';
 import { checkApiAvailability, editCharDebounced, saveAltGreetings } from '../services/characters-service.js';
 import { refreshCharListDebounced } from '../components/charactersList.js';
 import { updateTokenCount } from '../utils.js';
@@ -89,7 +88,7 @@ export function initializeCharactersEvents() {
     acm.st.eventSource.on('characterDeleted', function () {
         let charDetailsState = document.getElementById('char-details');
         if (charDetailsState.style.display !== 'none') {
-            closeDetails();
+            acm.eventManager.emit('acm_closeDetails');
         }
         refreshCharListDebounced(true);
     });
