+++ scripts/events/global-events.js
@@ -1,4 +1,3 @@
-import { initializeModalEvents, initializeUIMenuEvents } from './modal-events.js';
 import { initializeCharactersEvents } from './characters-events.js';
 import { initializeCharactersListEvents, initializeToolbarEvents } from './charactersList-events.js';
 import { initializeCharacterCreationEvents } from './characterCreation-events.js';
@@ -9,9 +8,7 @@ import { initializeCharacterCreationEvents } from './characterCreation-events.js
  */
 export function initializeEventHandlers() {
     // Initialize event handlers for different components
-    initializeModalEvents();
     initializeToolbarEvents();
-    initializeUIMenuEvents();
     initializeCharactersListEvents();
     initializeCharactersEvents();
     initializeCharacterCreationEvents();
