+++ templates/modal.html
@@ -476,7 +476,7 @@
                 <span id="separator-label">Advanced Definitions</span>
                 <span class="arrow arrow-up">&#x2193;</span>
             </div>
-            <div id="acm_right_panel" class="panel-transition">
+            <div id="acm_right_panel" class="panel-transition panel-hidden">
                 <div>
                     <h3 class="margin0"><span data-i18n="Advanced Defininitions">Advanced Definitions</span></h3>
                 </div>
